// src/api/generated/model/registerCompanyRequest.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { RegisterCompanyRequestCompanySize } from './registerCompanyRequestCompanySize';

export interface RegisterCompanyRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  street: string;
  /** @minLength 1 */
  zip: string;
  /** @minLength 1 */
  city: string;
  /**
   * @minLength 1
   * @pattern [A-Z]{2}
   */
  country: string;
  industry?: string;
  companySize?: RegisterCompanyRequestCompanySize;
  website?: string;
  /** @minLength 1 */
  adminEmail: string;
}